+++ app/Api/Application/Post/Services/DeletePost.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+
+namespace App\Api\Application\Post\Services;
+
+
+use App\Api\Domain\Models\Post;
+
+class DeletePost
+{
+    public function execute(Post $post): void
+    {
+        $post->delete();
+    }
+}
