+++ app/Api/Domain/Models/Comment.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace App\Api\Domain\Models;
+
+use App\Api\Infrastructure\Factories\CommentFactory;
+use App\Models\User;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+class Comment extends Model
+{
+    /** @use HasFactory<CommentFactory> */
+    use HasFactory;
+
+    protected $fillable = [
+        'post_id',
+        'user_id',
+        'content',
+        'rating',
+    ];
+
+    protected $attributes = [
+        'rating' => 0,
+    ];
+
+    protected $casts = [
+        'created_at' => 'datetime',
+        'updated_at' => 'datetime',
+    ];
+
+    protected static function newFactory(): CommentFactory|Factory
+    {
+        return CommentFactory::new();
+    }
+
+    public function post(): BelongsTo
+    {
+        return $this->belongsTo(Post::class);
+    }
+
+    public function user(): BelongsTo
+    {
+        return $this->belongsTo(User::class);
+    }
+
+    public function votes()
+    {
+        return $this->hasMany(CommentVote::class);
+    }
+}
