+++ app/Api/Domain/Policies/CommentPolicy.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Api\Domain\Policies;
+
+use App\Api\Domain\Models\Comment;
+use App\Api\Domain\Models\Post;
+use App\Models\User;
+
+class CommentPolicy
+{
+    public function update(User $user, Post $post, Comment $comment): bool
+    {
+        return $user->id === $comment->user_id;
+    }
+
+    public function delete(User $user, Post $post, Comment $comment): bool
+    {
+        return $user->id === $comment->user_id;
+    }
+}
