+++ app/Api/Domain/Policies/PostPolicy.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Api\Domain\Policies;
+
+use App\Api\Domain\Models\Post;
+use App\Models\User;
+
+class PostPolicy
+{
+    public function update(User $user, Post $post): bool
+    {
+        return $user->id === $post->user_id;
+    }
+
+    public function delete(User $user, Post $post): bool
+    {
+        return $user->id === $post->user_id;
+    }
+}
