--- app/Api/Enums/VoteDirection.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-
-namespace App\Api\Enums;
-
-
-enum VoteDirection: string
-{
-    case UP = 'up';
-    case DOWN = 'down';
-}
