--- app/Api/Events/CommentVoteDeleted.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App\Api\Events;
-
-use App\Api\Models\CommentVote;
-use Illuminate\Broadcasting\InteractsWithSockets;
-use Illuminate\Foundation\Events\Dispatchable;
-use Illuminate\Queue\SerializesModels;
-
-class CommentVoteDeleted
-{
-    use Dispatchable, InteractsWithSockets, SerializesModels;
-
-    public function __construct(public CommentVote $commentVote)
-    {
-    }
-}
