--- app/Api/Exceptions/ApiErrorHandler.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-
-namespace App\Api\Exceptions;
-
-
-use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
-use Throwable;
-
-class ApiErrorHandler
-{
-    public function __invoke(Throwable $e, Request $request): ?JsonResponse
-    {
-        if (!$request->is('api/*')) {
-            return null;
-        }
-
-        return (new ApiProblem($e))->toResponse();
-    }
-}
