--- app/Api/Factory/CommentVoteFactory.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace App\Api\Factory;
-
-use App\Api\Enums\VoteDirection;
-use App\Api\Models\Comment;
-use App\Api\Models\CommentVote;
-use App\Models\User;
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Api\Models\CommentVote>
- */
-class CommentVoteFactory extends Factory
-{
-    protected $model = CommentVote::class;
-
-    public function definition(): array
-    {
-        return [
-            'user_id' => User::factory(),
-            'comment_id' => Comment::factory(),
-            'direction' => $this->faker->randomElement([
-                VoteDirection::UP,
-                VoteDirection::DOWN,
-            ]),
-        ];
-    }
-}
