--- app/Api/Http/Controllers/PostController.php
@@ -1,69 +1,0 @@
-<?php
-
-namespace App\Api\Http\Controllers;
-
-use App\Api\Http\Resources\PostResource;
-use App\Api\Models\Post;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-
-class PostController extends Controller
-{
-    public function index()
-    {
-        return PostResource::collection(Post::paginate(10));
-    }
-
-    public function store(Request $request)
-    {
-        if (!$request->user()) {
-            abort(401);
-        }
-
-        $validated = $request->validate([
-            'title' => 'required|string|max:255',
-            'content' => 'required|string',
-        ]);
-
-        $post = Post::create(array_merge($validated, ['user_id' => $request->user()->id]));
-
-        return new PostResource($post);
-    }
-
-    public function show(Post $post)
-    {
-        $post->load('comments');
-
-        return new PostResource($post);
-    }
-
-    public function update(Request $request, Post $post)
-    {
-        if (!$user = $request->user()) {
-            abort(401);
-        }
-        if (!$user->can('update', $post)) {
-            abort(403);
-        }
-
-        $validated = $request->validate([
-            'title' => 'sometimes|required|string|max:255',
-            'content' => 'sometimes|required|string',
-        ]);
-
-        $post->update($validated);
-
-        return new PostResource($post);
-    }
-
-    public function destroy(Request $request, Post $post)
-    {
-        if (!$request->user()?->can('delete', $post)) {
-            abort(403);
-        }
-
-        $post->delete();
-
-        return response()->json(null, 204);
-    }
-}
