--- app/Api/Http/Resources/CommentResource.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace App\Api\Http\Resources;
-
-use App\Api\Models\Post;
-use App\Models\User;
-use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\JsonResource;
-use Illuminate\Support\Carbon;
-
-/**
- * @property int $id
- * @property string $content
- * @property int $rating
- * @property Carbon $created_at
- * @property Post $post
- * @property User $user
- */
-class CommentResource extends JsonResource
-{
-    /**
-     * Transform the resource into an array.
-     *
-     * @return array<string, mixed>
-     */
-    public function toArray(Request $request): array
-    {
-        return [
-            'id' => $this->id,
-            'content' => $this->content,
-            'rating' => $this->rating,
-            'created_at' => $this->created_at->toDateTimeString(),
-            'post_id' => $this->post->id,
-            'user' => [
-                'id' => $this->user->id,
-                'name' => $this->user->name,
-                'email' => $this->user->email,
-            ]
-        ];
-    }
-}
