--- app/Api/Http/Resources/PostResource.php
@@ -1,44 +1,0 @@
-<?php
-
-namespace App\Api\Http\Resources;
-
-use App\Api\Models\Comment;
-use App\Models\User;
-use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\JsonResource;
-use Illuminate\Support\Carbon;
-use Illuminate\Support\Collection;
-
-/**
- * @property int $id
- * @property string $title
- * @property string $content
- * @property Carbon $created_at
- * @property Collection<Comment> $comments
- * @property User $user
- */
-class PostResource extends JsonResource
-{
-    /**
-     * Transform the resource into an array.
-     *
-     * @return array<string, mixed>
-     */
-    public function toArray(Request $request): array
-    {
-        return [
-            'id' => $this->id,
-            'title' => $this->title,
-            'content' => $this->content,
-            'created_at' => $this->created_at->toDateTimeString(),
-            'comments' => CommentResource::collection(
-                $this->whenLoaded('comments', fn() => $this->comments->sortByDesc('rating'))
-            ),
-            'user' => [
-                'id' => $this->user->id,
-                'name' => $this->user->name,
-                'email' => $this->user->email,
-            ],
-        ];
-    }
-}
