--- app/Api/Models/Post.php
@@ -1,43 +1,0 @@
-<?php
-
-namespace App\Api\Models;
-
-use App\Api\Factory\PostFactory;
-use App\Models\User;
-use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\HasMany;
-
-class Post extends Model
-{
-    /** @use HasFactory<PostFactory> */
-    use HasFactory;
-
-    protected $fillable = [
-        'title',
-        'content',
-        'user_id',
-    ];
-
-    protected $casts = [
-        'created_at' => 'datetime',
-        'updated_at' => 'datetime',
-    ];
-
-    protected static function newFactory(): PostFactory|Factory
-    {
-        return PostFactory::new();
-    }
-
-    public function user(): BelongsTo
-    {
-        return $this->belongsTo(User::class);
-    }
-
-    public function comments(): HasMany
-    {
-        return $this->hasMany(Comment::class);
-    }
-}
