--- app/Api/Policies/PostPolicy.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Api\Policies;
-
-use App\Api\Models\Post;
-use App\Models\User;
-
-class PostPolicy
-{
-    public function update(User $user, Post $post): bool
-    {
-        return $user->id === $post->user_id;
-    }
-
-    public function delete(User $user, Post $post): bool
-    {
-        return $user->id === $post->user_id;
-    }
-}
