+++ app/Api/Shared/Enums/VoteDirection.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+
+namespace App\Api\Shared\Enums;
+
+
+enum VoteDirection: string
+{
+    case UP = 'up';
+    case DOWN = 'down';
+}
